* Renomeado namespace NFe.Classes.Informacoes.Detalhe.Produto Especifico para NFe.Classes.Informacoes.Detalhe.ProdEspecifico, para resolver problemas com espaços no VS2013
