+++ NFe.Classes/Informacoes/Detalhe/ProdEspecifico/CIDE.cs
@@ -30,7 +30,7 @@
 /* http://www.zeusautomacao.com.br/                                             */
 /* Rua Comendador Francisco josé da Cunha, 111 - Itabaiana - SE - 49500-000     */
 /********************************************************************************/
-namespace NFe.Classes.Informacoes.Detalhe.Produto_Especifico
+namespace NFe.Classes.Informacoes.Detalhe.ProdEspecifico
 {
     public class CIDE
     {
