+++ NFe.Classes/Informacoes/Detalhe/ProdEspecifico/ProdutoEspecifico.cs
@@ -30,29 +30,9 @@
 /* http://www.zeusautomacao.com.br/                                             */
 /* Rua Comendador Francisco josé da Cunha, 111 - Itabaiana - SE - 49500-000     */
 /********************************************************************************/
-namespace NFe.Classes.Informacoes.Detalhe.Produto_Especifico
+namespace NFe.Classes.Informacoes.Detalhe.ProdEspecifico
 {
-    public class arma : ProdutoEspecifico
+    public abstract class ProdutoEspecifico
     {
-        /// <summary>
-        ///     L02 - Indicador do tipo de arma de fogo
-        /// </summary>
-        public TipoArma tpArma { get; set; }
-
-        /// <summary>
-        ///     L03 - Número de série da arma
-        /// </summary>
-        public string nSerie { get; set; }
-
-        /// <summary>
-        ///     L04 - Número de série do cano
-        /// </summary>
-        public string nCano { get; set; }
-
-        /// <summary>
-        ///     L05 - Descrição completa da arma, compreendendo: calibre, marca, capacidade, tipo de funcionamento, comprimento e
-        ///     demais elementos que permitam a sua perfeita identificação.
-        /// </summary>
-        public string descr { get; set; }
     }
 }
