+++ NFe.Classes/Informacoes/Detalhe/ProdEspecifico/med.cs
@@ -30,64 +30,44 @@
 /* http://www.zeusautomacao.com.br/                                             */
 /* Rua Comendador Francisco josé da Cunha, 111 - Itabaiana - SE - 49500-000     */
 /********************************************************************************/
-namespace NFe.Classes.Informacoes.Detalhe.Produto_Especifico
+namespace NFe.Classes.Informacoes.Detalhe.ProdEspecifico
 {
-    public class comb : ProdutoEspecifico
+    public class med : ProdutoEspecifico
     {
-        private decimal? _pMixGn;
-        private decimal? _qTemp;
+        private decimal _qLote;
+        private decimal _vPmc;
 
         /// <summary>
-        ///     LA02 - Código de produto da ANP
+        ///     K02 - Número do Lote de medicamentos ou de matérias-primas farmacêuticas
         /// </summary>
-        public string cProdANP { get; set; }
+        public string nLote { get; set; }
 
         /// <summary>
-        ///     LA03 - Percentual de Gás Natural para o produto GLP (cProdANP=210203001)
+        ///     K03 - Quantidade de produto no Lote de medicamentos ou de matérias-primas farmacêuticas
         /// </summary>
-        public decimal? pMixGN
+        public decimal qLote
         {
-            get { return _pMixGn.Arredondar(4); }
-            set { _pMixGn = value.Arredondar(4); }
+            get { return _qLote; }
+            set { _qLote = value.Arredondar(3); }
         }
 
         /// <summary>
-        ///     LA04 - Código de autorização / registro do CODIF
+        ///     K04 - Data de fabricação. Formato: “AAAA-MM-DD”
         /// </summary>
-        public string CODIF { get; set; }
+        public string dFab { get; set; }
 
         /// <summary>
-        ///     LA05 - Quantidade de combustível faturada à temperatura ambiente
+        ///     K05 - Data de validade. Formato: “AAAA-MM-DD”
         /// </summary>
-        public decimal? qTemp
+        public string dVal { get; set; }
+
+        /// <summary>
+        ///     K06 - Preço máximo consumidor
+        /// </summary>
+        public decimal vPMC
         {
-            get { return _qTemp.Arredondar(4); }
-            set { _qTemp = value.Arredondar(4); }
-        }
-
-        /// <summary>
-        ///     LA06 - Sigla da UF de consumo
-        /// </summary>
-        public string UFCons { get; set; }
-
-        /// <summary>
-        ///     LA07 - Informações da CIDE
-        /// </summary>
-        public CIDE CIDE { get; set; }
-
-        /// <summary>
-        /// LA11 - Informações do grupo de “encerrante”
-        /// </summary>
-        public encerrante encerrante { get; set; }
-
-        public bool ShouldSerializepMixGN()
-        {
-            return pMixGN.HasValue;
-        }
-
-        public bool ShouldSerializeqTemp()
-        {
-            return qTemp.HasValue;
+            get { return _vPmc; }
+            set { _vPmc = value.Arredondar(2); }
         }
     }
 }
